pyproject.toml
[project]
name = "finance-inspector"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = ["pdfplumber", "python-dateutil", "typer", "rich"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
